// packages/contracts/generated/StartRunRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Id } from "./Id";
import type { RunContext } from "./RunContext";
import type { RunIntent } from "./RunIntent";

export type StartRunRequest = { operation_id: Id, profile_id: Id, context: RunContext, intent: RunIntent, prompt: string, consent_id: Id | null, };